Tagged forms and opaque values, and a refusal Rust was missing
`new-tagged`, `is-tagged`, `new-opaque`, `is-opaque`, `opaque-host-id` and
`opaque-label` from `kin/opaque.kin`. Six functions that lived on `Rt` in both
ports and in `coll.rs` in Rust.

A REFUSAL RUST DID NOT HAVE. Both ports checked `alloc` for failure in BOTH
constructors; Rust checked it in `new_opaque` and not in `new_tagged`. One
function, one runtime, and a half-allocated tagged form on a heap that had
already said no. `new-obj` answers NIL when the heap refuses, so using it makes
the check impossible to omit -- and the fixture starves the heap and watches
both answer nil, which is the branch none of the three had a test for.

THE HOST ID IS SIGNED NOW. Rust said `u64`, both ports `long`; the bits were
always the same and only the name for them differed. Three call sites in Rust
gained a cast, and the wire encoder says why it still writes unsigned.

`0022` IN ONE PLACE. An opaque value IS its identity: the id is STORED rather
than derived from the address, because the nursery is a copying collector and
an address-derived hash would change under collection -- a value in a map would
stop being findable by the key that put it there. The fixture mints two
opaques of the SAME label and asserts they are distinct.

AND `cargo check --all-targets` IS NOT THE WORKSPACE ORACLE. Run inside
`runtime/` it said zero errors while `flint-conc`'s host-ports test still used
the old `u64` signature; `bin/test` caught it. `cargo check --workspace
--all-targets` is the one that would have.

Gates: bin/test 0, conform-hosts 0 FAIL, cargo test 137 passed, workspace clean.

+++ runtime/src/codec.rs
@@ -234,7 +234,9 @@ impl Rt {
                 let mut b = crate::rt::sbuf();
                 let l: String = self.as_str(label, &mut b).unwrap_or("").into();
                 out.push(K_SENTINEL);
-                put_u64(out, self.opaque_host_id(v));
+                // SIGNED across the port now, and the wire is unsigned. The
+                // bits are the same; only the name for them differs.
+                put_u64(out, self.opaque_host_id(v) as u64);
                 put_str(out, &l);
                 Ok(())
             }
@@ -573,7 +575,7 @@ impl Rt {
                 let base = self.mark();
                 let l = self.string(&label);
                 self.push(l);
-                let out = self.new_opaque(self.r(base), host_id);
+                let out = self.new_opaque(self.r(base), host_id as i64);
                 self.pop_to(base);
                 Ok(out)
             }
